requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0

# LangGraph for agent orchestration
langgraph>=0.0.20

# Vector database backends (choose one or both)
chromadb>=0.4.0  # For local vector storage
pinecone-client>=2.2.0  # For cloud vector storage

# ML model persistence
joblib>=1.3.0

# Logging and utilities
python-dotenv>=1.0.0
google-generativeai>=0.8.0

# Optional: for enhanced features
# sentence-transformers>=2.2.0  # For better embeddings
# faiss-cpu>=1.7.0  # Alternative vector search

# API
flask>=3.0.0
flask-cors>=4.0.0